+++ app/src/main/java/dev/sergei/miniwebserver/data/AndroidAppInfoProvider.kt
@@ -1,0 +1,11 @@
+package dev.sergei.miniwebserver.data
+
+import dev.sergei.miniwebserver.BuildConfig
+import dev.sergei.miniwebserver.domain.app.AppInfoProvider
+import javax.inject.Inject
+
+class AndroidAppInfoProvider
+    @Inject
+    constructor() : AppInfoProvider {
+        override fun version(): String = BuildConfig.VERSION_NAME
+    }